condição para quando for pagina registro, trazer form de registro
