+++ src/pages/pickOrder/index.js
@@ -1,6 +1,8 @@
 /* eslint-disable no-console */
 import React, { useEffect, useState } from 'react';
+import Button from '../../components/Button';
 import FaceCapture from '../../components/FaceCapture/FaceCapture';
+import Header from '../../components/Header/Header';
 
 const PickOrder = () => {
   const [imgSrc, setImgSrc] = useState();
@@ -78,8 +80,11 @@ const PickOrder = () => {
 
   return (
     <>
+      <Header/>
       <FaceCapture label="Pegar Encomenda" setImgSrc={setImgSrc} />
-      <h1>PickOrder</h1>;
+      <Button
+        buttonText="Pegar pelo Código"
+      />
     </>
   );
 };
